플레이어 움직임 업데이트
