+++ UserSettings/EditorUserSettings.asset
@@ -9,7 +9,16 @@ EditorUserSettings:
       value: 5a5757560101590a5d0c0e24427b5d44434e4c7a7b7a23677f2b4565b7b5353a
       flags: 0
     RecentlyUsedSceneGuid-1:
+      value: 0150075652045d0e5f595b2345725944124f497d78797f632c704e35b0e1653b
+      flags: 0
+    RecentlyUsedSceneGuid-2:
+      value: 0508565751005b0e580f0a7311750f4410151d28797e71647f2a4f32e7e33139
+      flags: 0
+    RecentlyUsedSceneGuid-3:
       value: 5a52560401030b0b5d0a0e7148730844464f1b297a2b20637a2d4d35e7e6326e
+      flags: 0
+    RecentlyUsedSceneGuid-4:
+      value: 5a520d5e04015d580b57097141255e4446164a2e2d2a22312b794c30e7b43269
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
